requirements.txt
streamlit==1.29.0
google-genai==0.2.0
google-auth==2.23.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
pandas==2.1.0
plotly==5.18.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
schedule==1.2.0